Allow for nested case statements
The problem with nested case statements was that each branch of the case
statement (when, then, else, close) is responsible for consuming the
operations behind it to resolve the previous branch. E.g., when you hit
THEN you consume the operations before it to build the WHEN operation
and so on. This gets really messy when dealing with nested case
statements. There is probably a way to do this using the existing
approach but an easier way to think about it is that you can implode
each nested case statement and parse it by itself and just return the
output to the higher-level case statement.