Make register check constant only on Dentaku::AST
Currently, if there is a class called "Something" and a function
is registered with the name "something", a NameError will be raised.

That happens because Dentaku::AST::Function.register will try to verify
if "Something" is currently defined on Dentaku::AST with
"const_defined?".

If there is a class defined in the Top-level object with that name,
"const_defined?" will return true, and then the "register" method will
try to undefine it before redefining, throwing the NameError (as
Dentaku::AST::Something is not defined).

This fixes that error by adding the second parameter to the
"const_defined?" call as described here
https://ruby-doc.org/core-1.9.2/Module.html#method-i-const_defined-3F.

By calling "const_defined?" with false as the second parameter, it will
only check the constants defined in Dentaku::AST and not any of its
ancestors.