Arithmetic: speed up node validation
On large expressions, `node.dependencies` can take up a lot of time. On
some conditions, parsing a string can spend approximately all its time
traversing the dependency tree, validating if the left/right nodes are
ok.

We can avoid a lot of that time by first validating the type. This way
for example, for an expression like: '1 + 2 * 2', the Addition operator
does not need to traverse the Multiplication operator dependencies to
determine the node is indeed valid.

In order to simplify type checking, we add the type attribute to all
Node-like classes that don't inherit from Node.